+++ semantic-core/generation/semantics/db_name.py
@@ -1,0 +1,12 @@
+from typing import Annotated
+from pydantic import Field
+
+
+DbName = Annotated[
+    str,
+    Field(
+        description="The name of the database being connected to.",
+        examples=["customers"],
+        json_schema_extra={"is_sensitive": False},
+    ),
+]
