+++ semantic-core/generation/semantics/traces.py
@@ -1,0 +1,49 @@
+from typing_extensions import Annotated
+from pydantic import BaseModel, Field
+from typing import List, Dict
+
+
+""""
+This is for the payload structures only.  Fields do not belong here, outside of span/trace context fields
+"""
+
+
+SpanId = Annotated[
+    int,
+    Field(
+        description="""
+        Span identifier, generated by the tracer library. The value of this field is a 64-bit integer, and uniquely identifies the span within the trace.""",
+        examples=[0, 12345, 543210, 9999999],
+        json_schema_extra={"is_sensitive": False},
+    ),
+]
+
+TraceId = Annotated[
+    int,
+    Field(
+        description="""
+        Trace identifier, generated by the tracer library. The value of this field is a 64-bit integer, and uniquely identifies the trace within the org.""",
+        examples=[0, 12345, 543210, 9999999],
+        json_schema_extra={"is_sensitive": False},
+    ),
+]
+
+Tags = Annotated[
+    Dict,
+    Field(
+        description="""
+        This field represents an arbitrary map of key-value pairs.""",
+        examples=[{"foo": "bar", "key": "value"}],
+        json_schema_extra={"is_sensitive": False},
+    ),
+]
+
+TraceState = Annotated[
+    str,
+    Field(
+        description="""
+        Additional vendor-specific trace identification information across different distributed tracing systems. The tracestate field may contain any opaque value in any of the keys. See https://www.w3.org/TR/trace-context/#tracestate-header.""",
+        examples=["rojo=00f067aa0ba902b7", "rojo=00f067aa0ba902b7,congo=t61rcWkgMzE"],
+        json_schema_extra={"is_sensitive": False},
+    ),
+]
